+++ flutter_modular/lib/src/domain/services/bind_service.dart
@@ -2,7 +2,7 @@ import 'package:modular_core/modular_core.dart';
 import '../../shared/either.dart';
 
 abstract class BindService {
-  Either<ModularError, T> getBind<T extends Object>();
+  Either<ModularError, BindEntry<T>> getBind<T extends Object>();
   Either<ModularError, bool> disposeBind<T extends Object>();
   Either<ModularError, Unit> releaseScopedBinds();
 }