+++ flutter_modular/test/src/domain/usecases/get_bind_test.dart
@@ -2,6 +2,7 @@ import 'package:mocktail/mocktail.dart';
 import 'package:flutter_modular/src/shared/either.dart';
 import 'package:flutter_modular/src/domain/usecases/get_bind.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:modular_core/modular_core.dart';
 
 import '../../mocks/mocks.dart';
 
@@ -9,8 +10,8 @@ void main() {
   final service = BindServiceMock();
   final usecase = GetBindImpl(service);
   test('get bind', () {
-    when(() => service.getBind<String>()).thenReturn(right('test'));
+    when(() => service.getBind<String>()).thenReturn(right(BindEntry(bind: BindEmpty(), value: 'test')));
 
-    expect(usecase.call<String>().getOrElse((left) => ''), 'test');
+    expect(usecase.call<String>().map((r) => r.value).getOrElse((left) => ''), 'test');
   });
 }