+++ flutter_modular/test/src/infra/services/bind_service_impl_test.dart
@@ -13,13 +13,12 @@ void main() {
 
   group('getBind', () {
     test('should get bind', () {
-      when(() => injector.get<String>()).thenReturn('test');
-      expect(service.getBind<String>().getOrElse((left) => ''), 'test');
+      when(() => injector.get<String>()).thenReturn(BindEntry(bind: BindEmpty(), value: 'test'));
+      expect(service.getBind<String>().map((r) => r.value).getOrElse((left) => ''), 'test');
     });
     test('should throw error not found bind', () {
       when(() => injector.get<String>()).thenThrow(BindNotFound('String'));
-      expect(
-          service.getBind<String>().fold(id, id), isA<BindNotFoundException>());
+      expect(service.getBind<String>().fold(id, id), isA<BindNotFoundException>());
     });
   });
 
@@ -33,8 +32,7 @@ void main() {
   group('releaseScopedBinds', () {
     test('should return true', () {
       when(() => injector.removeScopedBinds());
-      expect(
-          service.releaseScopedBinds().getOrElse((left) => throw left), unit);
+      expect(service.releaseScopedBinds().getOrElse((left) => throw left), unit);
     });
   });
 }