+++ flutter_modular/test/src/presenter/modular_base_test.dart
@@ -54,8 +54,7 @@ class DisposableMock extends Mock implements Disposable {}
 
 class IModularNavigatorMock extends Mock implements IModularNavigator {}
 
-class ModularRouteInformationParserMock extends Mock
-    implements ModularRouteInformationParser {}
+class ModularRouteInformationParserMock extends Mock implements ModularRouteInformationParser {}
 
 class ModularRouterDelegateMock extends Mock implements ModularRouterDelegate {}
 
@@ -108,8 +107,7 @@ void main() {
     when(() => startModule.call(module)).thenReturn(right(unit));
     modularBase.init(module);
     verify(() => startModule.call(module));
-    expect(
-        () => modularBase.init(module), throwsA(isA<ModuleStartedException>()));
+    expect(() => modularBase.init(module), throwsA(isA<ModuleStartedException>()));
   });
 
   test('dispose', () {
@@ -123,19 +121,16 @@ void main() {
   });
 
   test('get', () {
-    when(() => getBind.call<String>()).thenReturn(right('modular'));
+    when(() => getBind.call<String>()).thenReturn(right(BindEntry(bind: BindEmpty(), value: 'modular')));
     expect(modularBase.get<String>(), 'modular');
   });
 
   test('getAsync', () {
-    when(() => getBind.call<Future<String>>())
-        .thenReturn(right(Future.value('modular')));
+    when(() => getBind.call<Future<String>>()).thenReturn(right(BindEntry(bind: BindEmpty(), value: Future.value('modular'))));
     expect(modularBase.getAsync<String>(), completion('modular'));
     reset(getBind);
-    when(() => getBind.call<Future<String>>())
-        .thenReturn(left(const BindNotFoundException('')));
-    expect(modularBase.getAsync<String>(defaultValue: 'changed'),
-        completion('changed'));
+    when(() => getBind.call<Future<String>>()).thenReturn(left(const BindNotFoundException('')));
+    expect(modularBase.getAsync<String>(defaultValue: 'changed'), completion('changed'));
   });
 
   test('isModuleReady', () {